Initial refactor of Omada message processing
Separating the code out to methods on the type for a cleaner implementation and being able to write better tests.

Also just couldn't decide where `BuildMessageBody` should live thanks to its old implementation, but with these newly added methods it looks like it is better to have it on the `gotify` package. (If we wanted to be *fancy* it should be using an interface; bit overkill for this program).